backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.38.0
google-generativeai>=0.8.5
supabase>=2.24.0
httpx>=0.28.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.2.1
anyio>=4.0.0
